requirements.txt
PyYAML~=6.0.2
Flask~=3.0.3
flask_socketio